Lab3 - Server sends arguments to client now

+++ src/FindNumber.java
@@ -60,13 +60,13 @@ public class FindNumber {
         return time;
     }
 
-    public static int findInRange(int a, int b, boolean print) {
+    public static int findInRange(long a, long b, boolean print) {
         int cnt = 0;
 
-        int from = Math.min(a, b);
-        int to = Math.max(a, b);
+        long from = Math.min(a, b);
+        long to = Math.max(a, b);
 
-        for (int i = from; i <= to; i++) {
+        for (long i = from; i <= to; i++) {
             if (i % 11 == 0 && i % 13 == 0 && i % 17 == 0) {
                 cnt++;
                 if (print)
